discordgo upgrade fixes

+++ discord.go
@@ -22,7 +22,7 @@ func messageToLower(message *discordgo.Message) *discordgo.Message {
 	return &newMessage
 }
 
-func hasPerms(channelID string, permission int) bool {
+func hasPerms(channelID string, permission int64) bool {
 	sourceChannel, err := bot.State.Channel(channelID)
 	if sourceChannel != nil && err == nil {
 		switch sourceChannel.Type {
